combine schema and params into one file
