+++ sketches/boids/src/boids.params.ui.ts
@@ -1,4 +1,4 @@
-import type { Config } from './boids.schema';
+import type { Config } from './boids.params';
 
 export class ParamsUI {
   private form: HTMLFormElement;
